refactor: 테스트 모두 통과
