+++ src/test/java/com/woowacamp/soolsool/acceptance/fixture/RestFixture.java
@@ -1,0 +1,6 @@
+package com.woowacamp.soolsool.acceptance.fixture;
+
+public abstract class RestFixture {
+
+    static final String BEARER = "Bearer ";
+}
