+++ src/test/java/com/woowacamp/soolsool/fake/FakePayClient.java
@@ -1,7 +1,8 @@
-package com.woowacamp.soolsool.core.payment.infra;
+package com.woowacamp.soolsool.fake;
 
 import com.woowacamp.soolsool.core.payment.dto.response.PayApproveResponse;
 import com.woowacamp.soolsool.core.payment.dto.response.PayReadyResponse;
+import com.woowacamp.soolsool.core.payment.infra.PayClient;
 import com.woowacamp.soolsool.core.receipt.domain.Receipt;
 import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Component;
